Moving LogHelper to its own header
