+++ src/banditpam.cpp
@@ -6,6 +6,7 @@
  *
  */
 #include "kmedoids_algorithm.hpp"
+#include "log_helper.hpp"
 
 #include <carma.h>
 #include <armadillo>
