+++ src/kmedoids_pywrapper.cpp
@@ -8,6 +8,7 @@
  */
 
 #include "kmedoids_algorithm.hpp"
+#include "log_helper.hpp"
 
 #include <carma.h>
 #include <armadillo>
