+++ src/main.cpp
@@ -10,6 +10,7 @@
  */
 
 #include "kmedoids_algorithm.hpp"
+#include "log_helper.hpp"
 
 #include <armadillo>
 #include <chrono>
